PROXY-7: address pull request comments
