--- controllers/asteroids
@@ -1,21 +1,0 @@
-const getAsteroidsWithinPeriod = require('../services/asteroidsData');
-const { startDate, endDate } = require('../helpers/dates');
-const transformAsteroidsData = require('../transformers/asteroids');
-
-const handleError = (res, error) => {
-    res.status(500).send(error.message);
-}
-
-const getAsteroidsData = async (req, res) => {
-    try {
-        const asteroidsData = await getAsteroidsWithinPeriod(startDate, endDate);
-        const transformedData = transformAsteroidsData(asteroidsData);
-        res.status(200).json(transformedData);
-    } catch (error) {
-        (error) => handleError(res, error);
-    }
-}
-
-module.exports = {
-    getAsteroidsData,
-};
