+++ package.json
@@ -4,7 +4,7 @@
   "description": "This is a project for learning Node/Express. In fact it will be proxy server for original NASA API.",
   "main": "index.js",
   "scripts": {
-    "dev": "node --watch -r dotenv/config ./server.js",
+    "dev": "node --watch -r dotenv/config ./src/server.js",
     "test": "echo \"Error: no test specified\" && exit 1",
     "eslint": "npx eslint ."
   },
