+++ src/routes/meteors.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { getAsteroidsData } = require('../controllers/asteroids');
+const { getAsteroidsData } = require('../controllers/asteroids.js');
 
 const router = express.Router();
